Rename module pre_conditioner -> preconditioner
