+++ src/pydvl/influence/torch/preconditioner.py
